move OptionaleTrait to Utils namespace
